Phase 1: scaffolding for cached-geometry refactor
- TiltProjection.ts: moved from game/boat/tessellation.ts (pure move; old
  location now re-exports).
- VertexSink.ts: sink interface + writeVertex helper with 7-float stride
  (position + color + z); transform/tint live per-instance, not per-vertex.
- TransformBuffer.ts: per-frame CPU/GPU storage for up to 4096 instance
  transforms (16 floats each: modelCol0..2 + zCoeffs + zDepth + tint). Slot
  0 is identity root; alloc returns 0 on overflow with a one-shot warning.
- CachedMesh.ts / DynamicMesh.ts: retained-mode mesh containers. Indices
  are 0-based and rebased at submission. No GPU resources held.
- curves.ts: pure geometry helpers (roundCorners, subdivideSmooth,
  subdivideCatmullRom, catmullRomOutputCount) moved out of boat/tessellation.ts;
  old location re-exports.

Nothing consumes the new files yet — this compiles clean but is inert
until Phase 3 wires them into the renderer.
